+++ src/openalea/distributed/cloud_infos/credentials.py
@@ -1,2 +1,7 @@
 # le meilleur endroit pour partager ses identifiants :)
 # IRODS
+irods_sess = iRODSSession(host='inragrid-mtp.supagro.inra.fr',
+                              port=1247,
+                              user='gheidsieck',
+                              password='ghe2018#',
+                              zone='INRAgrid')
